reorganize devices folder structure
