+++ pros/src/devices/adi/mod.rs
@@ -1,0 +1,42 @@
+/// Represents an ADI (three wire) port on a V5 Brain or V5 Three Wire Expander.
+#[derive(Debug, Eq, PartialEq)]
+pub struct AdiPort {
+    /// The index of the port (port number).
+    ///
+    /// Ports are indexed starting from 1.
+    pub index: u8,
+
+    /// The index of this port's associated [`AdiExpander`].
+    ///
+    /// If this port is not associated with an [`AdiExpander`] it should be set to `None`.
+    pub expander_index: Option<u8>,
+}
+
+impl AdiPort {
+    /// Create a new port.
+    ///
+    /// # Safety
+    ///
+    /// Creating new `ADIPort`s is inherently unsafe due to the possibility of constructing
+    /// more than one device on the same port index allowing multiple mutable references to
+    /// the same hardware device. Prefer using [`Peripherals`] to register devices if possible.
+    pub unsafe fn new(index: u8, expander_index: Option<u8>) -> Self {
+        Self {
+            index,
+            expander_index,
+        }
+    }
+
+    /// Get the index of the port (port number).
+    ///
+    /// Ports are indexed starting from 1.
+    pub fn index(&self) -> u8 {
+        self.index
+    }
+
+    /// Get the index of this port's associated [`AdiExpander`] smart port, or `None` if this port is not
+    /// associated with an expander.
+    pub fn expander_index(&self) -> Option<u8> {
+        self.expander_index
+    }
+}
