+++ pros/src/devices/battery.rs
