+++ pros/src/devices/mod.rs
@@ -1,0 +1,9 @@
+pub mod adi;
+pub mod smart;
+
+pub mod battery;
+pub mod controller;
+pub mod position;
+
+pub use position::Position;
+pub use controller::Controller;
