+++ pros/src/devices/smart/optical.rs
